Move extensions & bot settings under settings tab group
