+++ dev/settings/bot/settings.js
