+++ dev/settings/index.js
@@ -1,42 +1,5 @@
-const storage = require('libraries/storage');
-const STORAGE_ID = 'mb_preferences';
+const ui = require('ui');
+ui.addTabGroup('Settings', 'settings');
 
-var prefs = storage.getObject(STORAGE_ID, {}, false);
-
-// Auto save on change
-// IE (all) / Safari (< 10) doesn't support proxies
-if (typeof Proxy == 'undefined') {
-    module.exports = prefs;
-    setInterval(function() {
-        storage.set(STORAGE_ID, prefs, false);
-    }, 30 * 1000);
-} else {
-    module.exports = new Proxy(prefs, {
-        set: function(obj, prop, val) {
-            if (obj.hasOwnProperty(prop)) {
-                obj[prop] = val;
-                storage.set(STORAGE_ID, prefs, false);
-                return true;
-            }
-            return false;
-        }
-    });
-}
-
-var prefsMap = [
-    {type: 'number', key: 'announcementDelay', default: 10},
-    {type: 'number', key: 'maxResponses', default: 2},
-    {type: 'boolean', key: 'notify', default: true},
-    // Advanced
-    {type: 'boolean', key: 'disableTrim', default: false},
-    {type: 'boolean', key: 'regexTriggers', default: false},
-    {type: 'boolean', key: 'splitMessages', default: false},
-    {type: 'text', key: 'splitToken', default: '<split>'},
-];
-
-prefsMap.forEach(pref => {
-    // Set defaults if not set
-    if (typeof prefs[pref.key] !=  pref.type) {
-        prefs[pref.key] = pref.default;
-    }
-});
+require('./bot/page');
+require('./extensions');
