+++ dev/ui/layout/layout.html
@@ -13,7 +13,7 @@
 
 <div class="nav-slider-container">
     <nav class="nav-slider" data-tab-group="main"></nav>
-    <div class="overlay"></div>
+    <div class="is-overlay"></div>
 </div>
 
 <div id="container" class="has-fixed-nav"></div>
